Modify the type of the variables, This function that returns the number of bits you would need to flip to get from one number to another with bit manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,8 +9,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int flipbit;
-	unsigned int xorop;
+	unsigned long int flipbit;
+	unsigned long int xorop;
 
 	xorop = n ^ m;
